refactor(apis): http 메소드 별로 나뉘어져 있던 api로직 병합
